Integrate cli into existing i18n cli file
