+++ package.json
@@ -19,7 +19,7 @@
     "lint:tsc": "tsc --allowJs false --noEmit",
     "postinstall": "husky && patch-package",
     "translate": "node src/i18n/i18ncli.js build",
-    "prepare-metadata": "node fastlane/metadata/i18ncli.js build",
+    "prepare-metadata": "node src/i18n/i18ncli.js prepareFastlaneMetadataDirectories",
     "e2e:android": "npm run e2e:build:android && npm run e2e:test:android",
     "e2e:build:android": "MOCK_MODE=e2e npx detox build --configuration android.release",
     "e2e:build:ios": "MOCK_MODE=e2e npx detox build --configuration ios.release",
